[INTERNAL] sap.m.Select: visible second column value and tooltip
When read-only Select is diplayed in a two column layout now the value of
the Select includes the "text" and the "additionalText" of the selected
item, separated by a character, provided by SelectTwoColumnSeparator
enumeration. Possible values are Dash, Bullet and VerticalLine.

When Select is read-only (editable=false)
there is a default tooltip shown if text is truncated. The default
tooltip is the "text" of the item + "additionalText", if any, separated
by "-".

If the user has already set "tooltip" association, it has precedence
over the default tooltip. In this case, the default tooltip is not
provided.

JIRA: BGSOFUIPIRIN-6836
Change-Id: I40077e8e586bec91c5f80ab9c09d49800e5b2e27