[INTERNAL] mdc/Table: action ordering in the Toolbar
 - The table toolbar design guidelines describe several actions that are
not provided by the table itself but can be offered by applications or
Fiori Elements, such as the ViewSwitch or FullScreen button.
 - A new concept for tables and charts was required to enable additional
“standard” actions and ensure their correct positioning on the table
toolbar.
 - The IActionLayoutData marker interface defines the layout data used
to position actions within the toolbar and configure the overflow menu.
The layout data must extend sap/m/OverflowToolbarLayoutData and
implement a position property, which uses an enumeration to define the
relative order of actions. The sequence of the enumeration values
determines the positioning of actions. Enumeration values can be grouped
by using a shared prefix ending with the “Actions” keyword (e.g.,
ClipboardActionsCopy and ClipboardActionsPaste).
 - The new mdc/table/ActionLayoutData defines the layout data for
positioning toolbar actions within the mdc/Table according to the new
toolbar design guidelines.
 - The order of the enumeration values in mdc/enums/TableActionPosition
determines the order of actions.
 - The mdc/Table introduces a new tableActions aggregation that allows
adding table-related actions. These actions are positioned together with
the table-generated actions according to the provided ActionLayoutData.
While the actions aggregation supports UI adaptation, the tableActions
aggregation is excluded from it.
 - The ActionToolbar defines a new controlActions aggregation, which is
populated by the tableActions aggregation of the mdc/Table through
aggregation forwarding.
 - The getContent method of the ActionToolbar merges the end and
controlActions aggregations and sorts them according to the
ActionLayoutData of each action.
  - By using the group information from the position property, the
ActionsToolbar can generate group separators. This behavior is
explicitly excluded in the BLI; therefore, it is available only when the
sap-ui-action-toolbar-separators URL parameter is set.
 - The private _endOrder property of the ActionToolbar is no longer
needed. Related code and tests have been removed.
 - Demokit sample is added

Jira: CPOUIFTEAMB-2559
Jira: FIORITECHP1-33670

Change-Id: I634b209d5a24d3acfcfff1b9b64e62c64402a043