[INTERNAL] mdc.FilterBar: Implement new "Adapt Filters" design
Due to accessibility issues, a new design for the "Adapt Filters" panel has been proposed, which is now introduced with this change.
The new design is currently hidden behind a URL flag. To activate it add the URL parameter sap-ui-xx-new-adapt-filters=true to the URL.

This change introduces AdaptFilterPanelContent, a panel that implements the new design. AdaptFiltersPanel utilizes these panels.

PS 71:
- fix visibility action's enabled/disabled state for required fields
- adjust tooltips for OK + Cancel button to include shortcut hints
- adjust paddings and margins + background color for add filter section
- fix positioning
- adjust validation (move to AdaptationFilterBar)

PS73:
- adjust responsiveness

JIRA: CPOUIFDENVER-2174
Change-Id: I63226a492bf5e3e8d063a56ec267e5fa0587b3a4