[FIX] sap.uxap.ObjectPageLayout: Prevent focus loss when header expanded
Problem: In some cases, when header gets expanded with the Expand header
button, focus is lost (not moved to the Collapse header button).

Solution: This happens in very rare cases (on my machine only when a
certain app is opened on my laptop screen), when the scrollable wrapper
of the ObjectPage is scrolled a little bit further down, than the
snapping position + 1 px. For me it is snapping position + ~1.6px.
In this case, header moves to the title area, then moves back to the
content area. Because of the DOM tree rerendering focus gets lost
(focusout even is fired on the Collapse button).
That's why now, when we check if the page is scrolled to top, we round
down (floor) the scroll position of the scrollable wrapper.

SNOW: DINC0560683
Change-Id: Ib7dd57b5cd3989d30bf5809dfd376a99e8ed110e