[FEATURE] sap.f.ProductSwitchItem: Added imageSrc property
This commit introduces a new property `imageSrc` to the
`ProductSwitchItem` control in the SAP Fiori library.
This allows developers to use custom image for the product switch
item, enhancing the visual representation of items in
the product switch.

The new property takes precedence over the `src` property, meaning that if
`imageSrc` is set, the `src` property will not be rendered. This allows
for more flexible and visually appealing item representations.

JIRA: BGSOFUIPIRIN-6875
Change-Id: Ie3b4e507a9295d9e75f4476c81acb16ae68fb2c5