[INTERNAL][FIX] sap.ui.dt: DesignTimeMetadata - fix bundle handling (rereinstatement)
When retrieving the resource bundle from control libraries, we were
trying to load the bundle as if it were a library - which caused an
extra request for a manifest file that does not exist.

This change uses the proper method to create the bundle with its path.

This is a reinstatement of a previous fix that was reverted because the
in some cases the bundle was not available.

JIRA: TEAM42-415
Change-Id: I12d87e25869c52e66a267fd664c8a6f9b978ae20