[FIX] sap.m.OverflowToolbar: Fix failing test on 2.0
Problem: Due to different versions of jQuery, in some cases (when a
Control, placed in the Popover of the OverflowToolbar) gets its margins
measured in _getControlMargin function, in 2.0 NaN is returned instead
of 0. As a result aggregating controls to move to Popover is not
correct.

Solution: Check if Control is rendered and if not - return 0 for its
margins.

SNOW: DINC0592557
SNOW: DINC0592624
SNOW: DINC0592637
SNOW: DINC0592663
Change-Id: I91eac40394b15cefe5ee71d45342ef9bddb7b0eb