[INTERNAL][FIX] test/generic/Utils: Property Capabilities are now correctly interpreted per property
A "var" declaration caused a once set value for "sPropertyCapability" to be
taken over into the next loop.
This causes following properties to use the same capabilities as the one before.
This is especially bad if no exclude reasons should be taken into account,
as in this case a property that should be tested is accidentally skipped.

We now use a loop-local "let" declaration.

Additionally we exclude the respective properties from test execution.
In the future we might add a support for value set restrictions
for "int" based types in our generic testsuite.

Change-Id: I9d0430a00323d89e5234882cf7c7b77a104ad1e0