[FIX] FilterField: typeahed leads to error in phone mode
On phone mode typeahed opens in a full screen dialog. User enters some characters to have a filtering on the list. Now user selects some entries from the list. The typed characters stay in the input field.
If user presses OK the input is validated. If the types charaters itself are not a valid input (not a 100% match in the list or no other condition like ">10") an error is shown and the dialog cannot be closed.

Change-Id: I4a7af2d44fe0a01df3846eacab5ad1d32425813a
SNOW: DINC0505965