[FIX] sap.uxap.ObjectPageSection: Invalid aria-labelledBy fixed
Problem: When ObjectPageSection has only one SubSection and both their
titles are hidden (showTitle=false), Section still has aria-labelledBy
attribute pointing to the title of the promoted SubSection, which in
this case is not rendered and not visible. This is reported as invalid,
as there is no DOM element with such ID on the page.

Solution: When both titles of the Section and its promoted SubSection
(single one) are not visible, aria-labelledBy attribute is not set on
the Section.

SNOW: DINC0549818
Change-Id: Iad7a408fcb2fc4923b992d92558d96e0ce0df55e