[INTERNAL] m/Table: Accessibility improvements
 - The growing button lies outside the table’s keyboard navigation
because the table contains tabular content, while the growing button is
a standalone element. The GrowingEnablement event delegation moves focus
between the trigger and the last row when the Up/Down arrow keys are
pressed. This should only occur when the trigger is visible. The event
handler stops the immediate propagation of the event to prevent
ItemNavigation from handling it. In the reported case, this caused the
focus to become stuck.
SNOW: DINC0107877

 - When the table contains popins, a custom focus outline is used, which
requires calculating the first and last visible cells to correctly
display the left and right borders of the outline. If the focus is on a
row and a press handler triggers a re-render, the first and last cell
classes are removed from the DOM during rendering. This change ensures
that the handler is called after rendering to restore the first and last
classes when the focus remains on the same row.
SNOW: DINC0587081

 - For text selection purposes, tabindex="-1" remains necessary on the
popin row. The role="none" attribute is removed from the focusable popin
row. According to best practices, role="none" should only be applied to
purely decorative or structural elements and such elements must not be
focusable.
SNOW: DINC0535141

Change-Id: I6e0273a643724c7d46102a54e3dcee5c12713208