[INTERNAL][FIX] JSDoc: fix calculation of fully qualified names
When destructuring was used to create a shortcut for an enum type and
when the RHS of the destructuring was a member expression, not a simple
identifier, then the calculated name was incomplete and the enum was
not recognized properly, e.g. when used for a property default value.

Example:
  const {SimpleFormLayout} = layoutLibrary.form;

was mistakenly resolved to `sap.ui.layout.SimpleFormLayout` instead of
`sap.ui.layout.form.SimpleFormLayout`.

Change-Id: Ia741905776b7ab947d71b1325f7726497b53b83c