[INTERNAL]: sap.ui.fl - move cached backend response from FlexState to Loader
The Loader now manages the cached backend response for flexibility data.
This also includes checks if the cached data is still valid, e.g. by checking
the version or the allContexts flag.

The original change was reverted due to safetynet issues.
This new change has the following enhancements:
  - the skipLoadBundle parameter is only considered if it was previously set to true
  - an emptyCache gets invalidated as soon as getFlexData is called again
  - the skipLoadBundle is separated from the other parameters (e.g. version)

JIRA: TEAM42-464

Change-Id: I18c5b93ed2cd5a60e7357c35a2a20f46ad6e3c55