[FIX] sap.m.SinglePlanningCalendarMonthGrid: fixed Month view layout (320-340px)
This is complementary fix of change #6632487 for Compact mode.
The mentioned  previous fix covered only Cozy mode.

Issue: on desktop devices where the week number is displayed on Month
view, there was displacement of the days of week header and the dates
below because of the min-width set for day cells.

Solution: now the min-width is removed for smaller resolutions, and
the issue is fixed.

SNOW: DINC0172416
Change-Id: Ic1ff9b7afda34f250d579c08d964a72c1031bdd3