[INTERNAL]sap.ui.fl: Add init modules to apply and write bundles
The code contained in FL library.js should only contain initialization
for the 'initial' package. To ensure that the initialization logic for
'apply' and 'write' are always executed when needed, all publicly
available classes from the 'apply' and 'write' packages must reference
the preparation modules explicitly.

This change introduces an init module for both the 'apply' and 'write'
packages, which is required by all public classes in these packages.

JIRA: TEAM42-466
Change-Id: Ied1f59b67c3226123331edac71ad3243b1a4f39c