[FIX] sap.m.SegmentedButton: fixed initial truncation of some button texts
Issue: sometimes the text of some buttons inside the SegmentedButton
were truncated even there is enough space for the full texts. The
truncation disappears if there is some (even slight) resizing or
rerendering.

Cause: investigation of the problem found that the issue happens
because of the slow fonts loading, and because of the delay, the
initial sizes of the buttons are calculated before the complete fonts
loading, and is wrong. Later, when there is resizing/rerendering, the
sizes are recalculated again, this time with the proper text size.

Solution: the calculation and update of the widths of the buttons is
called in promise which resolves after the complete fonts loading
using the Fonts API.

Note: this fix covers case #1 in the BLI
Snippix: #995325

JIRA: BGSOFUIBALKAN-9029
Change-Id: I298d23c356299be1ec995b4cfdf597cb8fbf3cac