[INTERNAL] sap.ui.fl: Fix performance issue when updating FlexState
With a recent commit the update logic was changed and the internal FlexState cache
was not directly updated anymore, and instead was taken from the Loader cache.
With this an additional max layer filtering was necessary, which caused the performance impact.
Now the FlexState cache is again directly mutated, same as the Loader Cache.

- refactored the update logic to StorageUtils where it is used for both files
- renamed some methods so better reflect their purpose
	- FlexState.update -> FlexState.reinitialize
	- FlexState.updateStorageResponse -> FlexState.update
	- Loader.updateStorageResponse -> Loader.updateCachedResponse

SNOW: DINC0616338
Change-Id: I60399f8de7b6df98eafb7bf8804d13472ba5fe0a