[INTERNAL][FIX]sap.ui.fl: Fixed circular dependency issue in init module
Moving the change handler registration to the apply/init module caused
circular dependency issues because the iFrame change handler depends on
the IFrame util which in turn requires the fl library.

The dependency of the change handler on the IFrame is not needed; and we
also removed the dependency of the IFrame on the fl library.

SNOW: DINC0619886
Change-Id: I5551d1214268ddb2e37421de9b3bd04d1ae07d55