[INTERNAL][FIX]sap.ui.fl: Fix variant switch change apply/revert queuing
When switching variants, it is necessary to wait for the changes to be
applied or reverted before e.g. evaluating the App State. By setting
the changes to queued for apply/revert and adding the applied changes to
the dependency map, we can be sure they will all be finalized before
proceeding.

SNOW: DINC0601156
Change-Id: I5ea5def24a3aeb672399dd0b4503dc1aa3a187dc