[FIX] sap.ui.unified.Month: removed redundant "selected" announce
In the past the screen readed didn't announce the "selected" state
when a date was selected in the Month control, that's why there was
dedicated InvisibleMessage announcement of "selected" word.
Now screen reader already announces "selected" because of attribute
aria-selected, and the dedicated message causes double announcement.
That's why this announcement is removed now.

SNOW: DINC0514318
Change-Id: I456fb9178789f0c69a6d432c9c32521501c0ec21