[FIX] MDC.Chart: large numbers for measures are now correctly formatted
The Problem:
When large numbers which needs scientific notation for measures are provided, after formatting the value axis become invisible and the value in selection details become as question marks.

The Solution: Since there won't be adjustments of the formatting the solution for the value axis is by removing the formatter and for the value in the selection details we're not removing the formatter of the tooltip.

SNOW: DINC0477228
Change-Id: I20618a0feaec2864bff2bfd6da2b5536f3b39d9b