[FIX] sap.m.Avatar: Fix image not visible, when initials are incorrect
Problem: When Avatar has both incorrect initials (more than 3 latin
letters) and image src and the image is loaded later and Avatar is
rerendered (needed because there is also detailBox), image is not shown
- instead fallback icon is visible.

Solution: Problem happens because check for valid initials is always
done during rerendering. And if not valid - the actual type of Avatar is
changed to icon. However, this is not correct in all cases. Avatar needs
to fallback from initials to icon only if the actual display type is
initials. But if it is a valid src (image), it should not fallback to
icon.

SNOW: DINC0554089
Change-Id: I1af5a29863a16ca6ca4c6164d7164be3e71730e6