[FIX] AdaptFiltersPanel: Skip updateFinished logic, if in group view
If the SelectionPanel is showing the filter fields and has growing enabled, switching to the Group view and scrolling down lead to the filter fields vanishing from the group view.

This was due to the fact, that updateFinished (attached to the list in the Selection panel), still triggered, executing the logic to add the filter field to the selection panel list.

To avoid this, an inactive flag has been introduced, that can be set to "ignore" the logic in updateFinished.

SNOW: DINC0615996
Change-Id: Ic36cf0a0cb5068541ea659ad0ace5b088a32e2cd