[FIX] sap.uxap.ObjectPageLayout: Prevent duplicate roles on AnchorBar
Problem: The AnchorBar of the ObjectPage control contains wrapper
<section> element, which hosts IconTabHeader. By default, the
IconTabHeader has role="navigation". When the landmarkInfo aggregation
of the OP is used, the values of navigationRole and navigationLabel are
applied to the wrapper <section> controlled by the OP. This results in
two nested elements having the same of conflicting aria roles, depending
on the landmarkInfo configuration.

Solution: Apply the landmarkInfo values directly to the inner element
and avoid setting redundant role or label attributes on the wrapper
<section> element.

SNOW: DINC0628079
Change-Id: Iff93f2cedcf2479940004ecb30536623c1539671