[INTERNAL][FIX]sap.ui.dt: Ensure  focus is restored after closing menu
When the context menu is opened with mouse, the focus can be set to an
overlay not responsible for the menu (e.g. the label of a group element).
This can lead to issues with keyboard navigation after closing the menu,
as the popover tries to restore the focus to the last focused element
before opening (which can be a non-selectable overlay).

SNOW: DINC0622839
Change-Id: I12ee34d7de1adb821f2121b9312a09622cdfc9f8