[FIX] sap.ui.unified.CalendarAppointment: added 'ariaLabelledBy' association
Issue: there was no 'ariaLabelledBy' association in the
sap.ui.unified.CalendarAppointment control and when there is a case
of appointment with no title but with custom content, there must be
an option to add some other DOM element that will be read as label of
this appointment.

Solution: now 'ariaLabelledBy' association is being added to the
control, and when there is added element(s) that is used to label the
appointment, its text is being announced.

SNOW: DINC0446026
Change-Id: Id115111fbc4eec509d9d07778ed87d6a1498d3f0