[INTERNAL] sap.uxap.ObjectPageSection - hidden title aria-labelledby
Problem: After redesign of ObjectPageSections and SubSections, Sections
with hidden titles does not have aria-labelledby, as their title (now
Title control) is hidden (not rendered).

Solution: In this case, now we set the id of the anchorbar's button,
corresponding to the Section with hidden title, to the aria-labelledby
attribute of the Section.

JIRA: BGSOFUIPIRIN-6904
Change-Id: I2ceab31513d757b1c866f68df302267ed302b518