[FIX] sap.f.DynamicPage: Header visibility issue fixed
Problem:
The user cannot expand back the snapped header following
page-content resize. This happens if (1) with the new size the page
no longer shows a scrollbar (because all content fits the
new container size => header cannot be expanded by scrolling)
and (2) the app has disabled "toggleHeaderOnTitleClick".

This case is similar to the one fixed in Change-Id: Ie8c63110b8b598851b633ed1efd7bc418167a97f.
The difference is that here the resize is triggered
not by a change in the DynamicPage’s outer DOM element height,
but by a change in the height of its internal DOM element
containing nested application content.
In both cases, the resize makes the page non-scrollable
after it was previously scrollable.

Solution:
Auto-expand the snapped header on resize from scrollable
to non-scrollable.

SNOW: CS20250010602853
Change-Id: Ic1bbdb6abd89d8003641a0b1de0fb30d6efe412f