[INTERNAL][FIX]sap.ui.fl: Fixes for applying Annotation Changes
If the value to be set in a ChangeAnnotation change contains characters
that need to be escaped in a JSON string (e.g. double quotes),
the change application fails due to a parsing error. Now the parsing is
done before replacing the placeholder in the template string.

This change also ensures that if an annotation change fails, the
subsequent annotation changes are still applied.

SNOW: CS20250010499865
Change-Id: Iee0d1dc459dd2c7c23c60f0d85887c041f332c80