[INTERNAL] sap.ui.fl: initialize empty flex state with component ID
The component Id is necessary for some of the data in the FlexState, but
was not available for scenarios where an empty state is initialized.
This is the case when there is no state, but something should be added to it.

With not always initializing a complete FlexState on app start this will get
more relevant.

JIRA: TEAM42-475
Change-Id: Ia1e7ac43cebb235a6444959f9c2bf662a0e414d9