[FIX] ManagedObjectModel: Use original binding's length if no filters are applied
If a filter is applied, the original bindings length was taken, when calling getLength on the MOM Binding. This leads to an incorrect count, as the filter has not been applied on the original binding.

Therefore, the distinction is being made, to use the outer/inner binding's length if a filter is applied.

SNOW: DINC0561592
Change-Id: I9b7d7bcaab099c3e10efd969d3e60052e7b31e34