[INTERNAL] sap.uxap.ObjectPageLayout: Update aria-labels and landmarks
Aria-roledescriptions of ObjectPage, Header, AnchorBar and Footer are
now removed. Their predefined texts are now concatenated with the
predefined aria-labels.

If ObjectPageAccessibleLandmarkInfo is used and relevant role is set,
aria-label for the particular element (ObjectPage, Header, AnchorBar,
Footer) is overriden by the aria-label, defined in the landmark info by
the user.

JIRA: BGSOFUIPIRIN-6919
Change-Id: Iccb088af7d9318dc496623fed353a807f5976b9f