[FIX] VariantManagement: Replace table MOM binding with cloned binding
The ManagedObjectModel's purpose is to represent the control state in read-only cases.
It has multiple issues in write/modification scenarios, such as filtering for example (incorrect data state during filtering).
To avoid this problem altogether in the VariantManagement, it'll use a "copy" of the original items binding (if one exists), to ensure correct states during operations.
If no binding is used (items are manually added), the ManagedObjectModel binding is used as a fallback.

SNOW: DINC0659765
Change-Id: Ie35342cee2ae715c263fa3973c961bc251b51d69