[FIX] SelectionPanel: (De)select all should ignore hidden items
Only items that are actually visible should be affected by the (de)select all.

This is the case if a filter is set by using the search field, but was not ensured, when
"Hide Description" or other toggles were used.
When filtering the p13nData, all filters are taken into account (by ORing the filter results).

Filters are now recursively evaluated to account for all cases.

SNOW: DINC0627088
Change-Id: Icb3dfb32c223fc5172d06cffc9fa99f7415f19a5