[INTERNAL] FESR activation via configuration
The configuration option sap-ui-fesr could be a boolean or
a URL to the FESR service. The activation was wrongly implemented
in jquery.sap.trace.js and the support plugin, only checking for
a boolean value.
If FESR will be deactivated, passport still must intercept open
to add passport headers, but must only set the header for non
CORS requests.

Change-Id: Ic545cd9a145ae10458a4e728793eb44711a55c5f