[INTERNAL][FIX] MultiSelectionPlugin: Detect by the length of the contexts when limit not reached
When the binding doesn't request the count, it still provides a calculated value for the binding length which might be greater than the limit. In that case the MultiSelection plugin expects that the limit will be reached. Once the binding provides the contexts, the MultiSelectionPlugin should check the actual length of the provided contexts and detect if limit was not reached to prevent scrolling the table.

SNOW: DINC0379549
Change-Id: I58b63605924729808d4a9db3a240e0656d53d5f8