[FIX] xConfig: Remove correct value
If a key has multiple entries (e.g. filterConditions) upon removing it the first element is always being removed.
This is due to the fact, that there is no value comparison when searching for the to be removed item.
To fix this issue, a deep comparison of the payload value is being made, to ensure that the correct item is being removed.

Fixes: https://github.com/SAP/openui5/issues/4277
SNOW: DINC0490163
Change-Id: Ie70b0bc63e01e3257008f1736842a85658cf2111