[INTERNAL] Field: formatter in binding are not supported
If a Formatter is used in Binding the Binding gets one-way. This means no parsing of user input is specified and no model update on user input is possible.

In the current implementation this could lead to differen problems when using user intercation or a value help. (e.g. keys might be set in wrong format.)

Change-Id: I109bf8e4ee01461502c06281e4ecf0217d19b589
SNOW: DINC0635374