[INTERNAL][FIX] AlignedFlowLayout: Ensure positive values for margin calculation
If the AFL is used in combination with the mdc.FilterBar, special styles are applied on the individual layout items.
One of these special CSS values is a negative margin-left (-0.25rem). This negative margin leads to some visual artefacting, when the layout is wrapping. For a specific width, the fields in the 2nd row are smaller than the other fields.
This is due to the fact, that the negative margin is also being applied to the last spacer. If we enforce that margin to be positive, this visual issue does not occur, as spacers will be aligned correctly.

SNOW: DINC0609071
Change-Id: I55848b98a995db92678c8124b10f9cd7d6ef94a3