[FIX] Field: error message not removed if previous value selected
If the user enters an invalid value and error message is shown. This error message is also available in the MessageModel.
If then the user selects an item from the valueHelp with the same key, but a different decription or payload, the error message will not be removed from the MessageModel.

Change-Id: Ib8ea88c884b2ed1c2e534169e037214ad30009c3
SNOW: CS20250010606880