[INTERNAL] LegacyFreeUI5 - check failing tests
The Problem:
As we recently introduced a sample for MDC Chart demonstrating usage with ChartJS it seems that the manifest.json is not compliant fully with the guidelines and linter throws an error.

The Solution:
As described in https://ui5.sap.com/#/topic/be0cf40f61184b358b5faedaec98b2da#loiobe0cf40f61184b358b5faedaec98b2da/section_manifest2 we set the async flag to true for rootView as for the rest of the samples in the library.

SNOW: BGSOFUISAKAR-699
SNOW: BGSOFUISAKAR-908
Change-Id: Ia6966ba0d953426fe8a514cb8d2e28875ccb71a5