[FIX] sap.uxap.ObjectPageLayout: prevent flickering on scroll
Problem: When there is a difference in height between snapped and
expanded title and user tries to scroll down just below the snapping
point of the header, header gets snapped/unsnapped, causing flickering.
Moreover, user is not able to scroll to the bottom of the page.

Solution: It seems that upon scrolling in this situation, after header
is snapped, scroll position is changed to point, that according to
calculations, header needs to be unsnapped. This is causing infinite
snapping/unsnapping upon scrolling.
That's why now ObjectPage restores the actual scroll position (when
scroll event is fired and header needs to be snapped - we call scrollTo
to the same scroll position). This prevents flickering
(snapping/unsnapping) and sustains the current state (snapped or
unsnapped).

SNOW: DINC0133318
Change-Id: Ib7a1c1f073f469c86ece91fa06190f658fdb0a46