[FIX] SelectionController: Check for key or item in getPropertySetterChanges
getPropertySetterChanges tries to identify the existing item, but only uses the 'name' property to do so. This works in MDC cases, as it exposes/uses the 'name' property in xConfig cases, but does not work in freestyle use cases, as 'name' is not defined in such cases.

This lead to issues with diffing, where state changes were fired, although the same state has been applied.

Fixes: https://github.com/UI5/openui5/issues/4221
Change-Id: I4388270700df63411f48e4d9aa8f6d2e7a51a5c1