[INTERNAL] sap.ui.commons.FileUploader: Detach commons.FileUploader from unified.FileUploader
There is a refactoring of the sap.ui.unified.FileUploader control,
and as the sap.ui.commons.FileUploader just extends unified control
with the same name, it is necessary to detach these two controls as
the refactoring heavily depends on other sap.m controls.

For this purpose, the current unified.FileUploader code and CSS is
being copied to commons.FileUploader and now it has its own code/CSS
instead of extending from another control.

JIRA: BGSOFUIBALKAN-9588
Change-Id: I139d04a97a7723a4febe4bf4ba711cfba8c75971