[FIX] sap.uxap.ObjectPageLayout: Fix scrolling when element is focused
Problem: When an element in ObjectPageSubSection is focused, but it is
currently hidden behind the sticky header of the ObjectPageSection,
browser does not scroll to it, as it considers the element is into
view (although overlapped by sticky content).

Solution: Scroll-padding-top of the scrollable container of the
ObjectPageLayout is adjusted, so that if the page is scrolled below the
top position of the currently selected Section (header is sticky), the
height of the Section's header is added to the current
scroll-padding-top (height of the title area of ObjectPageLayout).
This allows the Browser to scroll to the overlapped element.

SNOW: DINC0601089
Change-Id: I22840428140662eb5c317e6ea434a59b5cde5d3b