[INTERNAL] FESR: Improve E2E time calculation for interactions
Check for last rendering that happens for the ownercomponent of the
interaction. This time result in the E2E time. We still track all
requests to not break the idle time detection.
For the avarage Networktime calculation we now exclude requests not
reporting the fesrec header, otherwise we get wrong numbers.
The calculation of the navigation time is corrected also, as the
overlapping calculation was not working correctly.
The processing time can be only calculated if all requests are tracked.
Otherwise we report -1 as we cannot calculate the time correctly. The
time will be calculated with all timings withing the interaction
duration. If a request started within the interaction but ends after the
interaction end, we cut the time at the interaction end.
When introducing the fetch interceptor we can improve this again.
The 'legacy' E2E time will be reported in extension field 4 for
comparison reasons.

JIRA: CPOUI5FRAMEWORK-811
Change-Id: I7c73ed30ef9015e7c09c237d1800dade2653185d