[INTERNAL] JSDoc: improve support for "module:" syntax
When the JSDoc name for a ManagedObject class differs from its runtime
name (as specified in the extend call), then use the JSDoc name to
calculate the fully qualified names of generated documentation.
Otherwise, JSDoc can't associate the generated documentation with the
class.

Also enhances the check for a mismatch between documented base type and
technical base type. It no longer reports a future error when module:*
syntax is used for the documented base type.

Change-Id: If358d04e7960b524bbb79e89e380a20d80962ee1