[FIX] sap.m.Avatar: fix wrong display of icon
Problem: When initials are empty string and the src of Avatar is valid
icon - default icon is shown, instead of the user set icon.

Solution: Move the whole logic for setting private property
_bUseDefaultIcon in the setUseDefaultIcon method. It is called from
three places:
- onBeforeRendering (where we have unified state - src and initials
  properties are set)
- onImageLoad (as it comes later)
- onImageError (as it comes later)

Additionally, loading of image is done only when src is changed and this
logic is separated from _getActualDisplayTypeBySrc method.

SNOW: DINC0660810
Change-Id: I4a1ad01593dd1ed8a2221d3f6ce175398c639871