[INTERNAL][FIX]sap.ui.rta: Fix WhatsNew dialog test
When the WhatsNew dialog is opened, a flag is set in sessionStorage to prevent
the dialog from being shown again after a reload. The test was unstable on firefox and safari
because the spy on sessionStorage.setItem did not work as expected so it has been removed and replaced
by a direct check of sessionStorage which also boosts performance.

SNOW: DINC0664487
Change-Id: I88773ccfa415813bc60769bc1f5fe0deb83f51f2