[INTERNAL] sap.ui.fl (test): fix broken stub in VariantModel test (2.0)
Sinon's `stub.withArgs(...)` method returns a new instance for the
conditional stub. Therefore, the return value must not be used to chain
calls that should affect the original stub, they'll affect only the
conditional stub.

This showed up in legacy free tests as the `sap.ui.core.Renderer` was
not able to require something while the broken stub was active.

Change-Id: I2951f684fe634c5b59be89787e85b7231e00c057