[INTERNAL] sap.ui.fl: New approach to set iframe src
Previously the iframe always had the same src and was only updated via
replaceLocation calls to avoid history entries on src updates.
Due to various timing issues in case of rapid updates this lead to
various problems, especially on Safari.

The new approach goes back to setting the src directly but rerenders
the iframe DOM element on every update and sets the src before
rendering.
As a result the src is considered as the initial value and doesn't add
a history entry.

JIRA: TEAM42-488
Change-Id: I642d6a9eb8a3398000e535a6e5c247ebd6e45746