[INTERNAL] mdc: Adjust error message for disabled property validation
If propertyInfo validation is disabled (e.g. in FE), the following error is thrown: "Property Info validation is disabled for control XYZ. Migrate this control's propertyInfo to avoid breaking changes in the future."

The suffix in this message is shown, although it is not confirmed, that the propertyInfos of the control are incorrect. This leads to confusion with FE developers and developers using FE.
To avoid confusion, the suffix has been changed to merely state the exemption being removed in 2.0. Additionally, instead of throwing an error, we are logging a warning instead.
FE and others still need to migrate their propertyInfos, to ensure apps running in 2.0.

SNOW: CS20250010885189
SNOW: DINC0493158
Change-Id: Ideab83f9029060d8c7d1fa4ad5df52e9f27de199