[FIX] sap.uxap.ObjectPageSubSection: prevent infinite rerendering loop
Problem: When application has onAfterRendering delegate, attached to a
SubSection, and changes the title, infinite loop of rerendering happens,
due to new logic of invalidating the Section when ariaLabelledBy id is
set to the relevant button of the AnchorBar.

Solution: This invalidation all the time is actually not needed as the
Button ID is always the same. Now we invalidate only if there is a
change in the Button ID (most probably won't ever happen).
Additionally, setTitle setter of ObjectPageSectionBase (inherited by
both Section and SubSection) is enhanced, so that if same title is set
repeatedly no action is taken (omitting notfiying ObjectPageLayout
etc.).

SNOW: CS20250010928227
Change-Id: If620ea50d68050955a612cab9aeef6a673e5fd27