[INTERNAL] Table: Clear selection in TreeTable hiearchy maintenance sample
If a node is removed in the JSONModel, the selection needs to be cleared manually to avoid issues in combination with selection mode Single.
The sample uses selection mode MultiToggle, but users changed to Single in the sample editor and experienced an issue. So the selection is cleared always for the sake of clarity.

Change-Id: I02590cb544975055e50c4f150bbabefbe8f913ec
SNOW: CS20250010828239