[FIX] We've now removed the iframe content check while loading a PDF to display it
PDFViewer uses an iframe along with the browser’s PDF plugin to display PDFs. Previously, we validated the iframe’s inner content to determine whether to load the PDF or show a Loading Error IllustratedMessage.

Modern browsers no longer return iframe content for invalid URLs, causing the validation to fail incorrectly.

We have now removed this iframe content check. The validation is now based solely on the presence of the HTML <embed> tag. If the <embed> tag is missing, we display the Loading Error IllustratedMessage.

cSNOW: DINC0680978
Change-Id: I164d846c3212df32049ab8dfc1d8795e046694da