[FIX] mdc.Chart: IllustratedMessage button for changing chart type is no longer visible when there are no visible measures
The Problem: When the visible measures are less than the required
number for a selected chart type we show IllustratedMessage with
button that provides option to select a different chart type -
a chart type that is available.
However, when there are no visible measures, then no chart types
are available, hence the button seems to be doing nothing.

The solution:
- Now we hide the button of there are no chart types available.
- We also changed the error message for this scenario

SNOW: DINC0634290
Change-Id: I8a1a4825c8bda883ef74b6781ff0ce2f8430a1c6